Fix spec.PL for Perl 5.26 '.' removal from @INC
Perl 5.26 removes '.' from @INC, and this breaks the semantics
of <<do $file>> where $file was assumed to be loaded from '.',
but not explicitly stated, as 'do' is implemented using the same code
as "require", that is, assuming @INC traversal.

Inserting a leading "./" forces do to use a special-cased path
where @INC is not consulted.

Fixes RT#120697

Bug: https://rt.cpan.org/Ticket/Display.html?id=120697
Bug: https://bugs.gentoo.org/613584